requirements.txt
alembic==1.7.5
apparmor==3.1.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Authlib==0.15.5
autocommand==2.2.1
azure-common==1.1.28
azure-core==1.24.2
azure-identity==1.9.0
azure-mgmt-core==1.3.1
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.10.3
bcrypt==3.2.2
Beaker==1.11.0
beautifulsoup4==4.11.1
bidict==0.22.0
blinker==1.5
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
btrfsutil==5.19.1
CacheControl==0.12.11
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
cryptography==3.4.8
cupshelpers==1.0
Cython==0.29.32
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
distro==1.7.0
Django==4.1.3
django-crispy-forms==1.14.0
django-translated-fields==0.12.0
djangorestframework==3.14.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
docx==0.2.4
email-validator==1.3.0
eventlet==0.33.0
extras==1.0.0
fixtures==4.0.1
Flask==2.1.3
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-pymysql==0.2.3
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.2
gssapi==1.8.0
html5lib==1.1
httpagentparser==1.9.3
idna==2.10
importlib-metadata==4.9.0
inflect==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
jdatetime==4.1.0
Jinja2==3.1.2
jmespath==0.10.0
kavenegar==1.1.2
kazam==1.4.5
keyutils==0.6
ldap3==2.9
LibAppArmor==3.1.1
libtorrent==2.0.7
lit==14.0.6.dev0
louis==3.23.0
lxml==4.9.1
Mako==1.2.3
mallard-ducktype==1.0.2
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
meson==0.63.3
more-itertools==8.13.0
msal==1.18.0
msal-extensions==0.3.1
msgpack==1.0.4
msrest==0.7.1
mysql-connector==2.2.9
mysqlclient==2.1.1
names==0.3.0
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.23.5
oauthlib==3.2.0
optimus-manager==1.4
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
panda==0.3.1
pandas==1.5.2
paramiko==2.11.0
passlib==1.7.4
pbr==5.10.0
pep517==0.13.0
persepolis==3.2.0
pgadmin4==6.12
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
portalocker==2.5.1
progress==1.6
psutil==5.9.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
pydantic==1.10.2
Pygments==2.13.0
PyGObject==3.42.2
PyJWT==2.4.0
pymongo==4.3.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyodbc==4.0.34
pyOpenSSL==22.1.0
pyotp==2.6.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.2
pysmbc==1.0.23
python-dateutil==2.8.2
python-distutils-extra==2.39
python-docx==0.8.11
python-editor==1.0.4
python-engineio==4.3.3
python-socketio==5.7.1
pytz==2021.3
pyxdg==0.28
PyYAML==6.0
qrcode==7.3.1
rados==2.0.0
rbd==2.0.0
redis==4.4.0
reportlab==3.6.11
requests==2.25.1
requests-oauthlib==1.3.1
resolvelib==0.8.1
retrying==1.3.3
rgw==2.0.0
rich @ file:///build/python-rich/src/rich-12.6.0/dist/rich-12.6.0-py3-none-any.whl
s3transfer==0.5.2
setproctitle==1.3.1
shiboken6==6.3.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
speaklater==1.3
speaklater3==1.4
SQLAlchemy==1.4.39
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==235
team==1.0
tenacity==8.1.0
testtools==2.5.0
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.9.26
typing_extensions==4.3.0
ua-parser==0.15.0
udiskie==2.4.2
urllib3==1.26.12
user-agents==2.2.0
validate-pyproject==0.10.1
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
youtube-dl==2021.12.17
zipp==3.8.1
